Refactoring. Extract test logic in test. Delete .gradle, .idea files to ignore in git.
