+++ .gitignore
@@ -3,3 +3,6 @@
 out
 .gradle
 build
+gradle
+/gradlew
+/gradlew.bat
