--- .gradle/buildOutputCleanup/cache.properties
@@ -1,2 +1,0 @@
-#Wed Apr 04 00:40:12 MSK 2018
-gradle.version=4.6
